_bmad/brain-swarm/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
numpy>=2.0.0
nilearn>=0.10.0
nibabel>=5.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
scipy>=1.13.0
